Error handling added to delete, post, and both GET requests

+++ routers.js
@@ -10,7 +10,9 @@ router.get("/", async (req, res) => {
     res.status(200).json(posts);
   } catch (error) {
     console.log(error);
-    res.status(500).json({ message: "error" });
+    res
+      .status(500)
+      .json({ message: "The posts information could not be retrieved." });
   }
 });
 
@@ -18,22 +20,36 @@ router.get("/:id", async (req, res) => {
   const { id } = req.params;
   try {
     const post = await db.findById(id);
-    res.status(200).json(post);
+    if (post) {
+      res.status(200).json(post);
+    } else {
+      res
+        .status(404)
+        .json({ message: "The post with the specified ID does not exist." });
+    }
   } catch (error) {
     console.log(error);
-    res.status(500).json({ message: "error" });
+    res
+      .status(500)
+      .json({ message: "The post information could not be retrieved." });
   }
 });
 
 router.post("/", async (req, res) => {
   const post = req.body;
-  try {
-    const edit = await db.insert(post);
-    const newPost = await db.findById(edit.id);
-    res.status(200).json(newPost);
-  } catch (error) {
-    console.log(error);
-    res.send(500).json({ message: "error" });
+  if (!post.title || !post.contents) {
+    res
+      .status(400)
+      .json({ message: "Please provide title and contents for the post." });
+  } else {
+    try {
+      const edit = await db.insert(post);
+      const newPost = await db.findById(edit.id);
+      res.status(201).json(newPost);
+    } catch (error) {
+      console.log(error);
+      res.send(500).json({ message: "error" });
+    }
   }
 });
 
@@ -41,11 +57,21 @@ router.delete("/:id", async (req, res) => {
   const { id } = req.params;
   try {
     const post = await db.findById(id);
-    const deleted = await db.remove(id);
-    res.status(201).json(post);
+    if (post) {
+      const deleted = await db.remove(id);
+      if (deleted) {
+        res.status(201).json(post);
+      } else {
+        res.status(500).json({ message: "The post could not be removed" });
+      }
+    } else {
+      res
+        .status(404)
+        .json({ message: "The post with the specified ID does not exist." });
+    }
   } catch (error) {
     console.log(error);
-    res.send(500).json({ message: "error" });
+    res.send(500).json({ message: "Something went wrong with your request." });
   }
 });
